Fixing portal scripts adding additional folders new folder still not working in pyinstaller fixing legacy logic to address new portal_english backup various readme fixes

+++ src/legacy_backup_tools.py
@@ -13,6 +13,9 @@ def get_backup_cfg_path(file_tools,type):
 def get_backup_other_path(file_tools):
     return file_tools.get_full_basegame_path() + "\\resource\portal_english-portl-backup.txt"
 
+def get_backup_other_path_2(file_tools):
+    return file_tools.get_full_basegame_path() + "\\resource\portal_english_backup.txt"
+
 def get_backup_captions_text_path(file_tools):
     return file_tools.get_full_basegame_path() +  \
            "\\resource\closecaption_{}-portl-backup.txt".format(file_tools.language)
@@ -72,11 +75,15 @@ def restore_captions(file_tools):
             os.remove(backup_captions_text_path)
 
 def restore_portal(file_tools):
-    backup_other_path = get_backup_other_path(file_tools)
-    file_data = {"name": "portal","folder":"resource","localized": True}
-    if (os.path.isfile(backup_other_path)):
-        copyfile(backup_other_path, file_tools.get_basegame_english_other_path(file_data))
-        os.remove(backup_other_path)
+    file_data = {"name": "portal","folder":"resource","localized": True,"extension":"txt"}
+    actual_file = file_tools.get_basegame_english_other_path(file_data)
+    if not (os.path.isfile(actual_file)):
+        backup_other_path = get_backup_other_path(file_tools)
+        if (not os.path.isfile(backup_other_path)):
+            backup_other_path = get_backup_other_path_2(file_tools)
+        if (os.path.isfile(backup_other_path)):
+            copyfile(backup_other_path, file_tools.get_basegame_english_other_path(file_data))
+            os.remove(backup_other_path)
 def restore_backup(file_tools):
     restore_cfg(file_tools)
     restore_portal(file_tools)
